Multiple selection update and minor group creating and editing fix
- Now is possible to permanently delete or recover multiple notes at the same time inside the recicle bin

- Selecting a new color or editing the existing one of a group now updates the color preview icon

+++ lib/presentation/screens/skeleton.dart
@@ -9,6 +9,7 @@ import 'package:simple_notes/presentation/screens/providers/note_provider.dart';
 import 'package:simple_notes/presentation/screens/providers/recicle_bin_provider.dart';
 import 'package:simple_notes/presentation/screens/providers/task_provider.dart';
 import 'package:simple_notes/presentation/screens/tasks_screen.dart';
+import 'package:simple_notes/presentation/widgets/dialogs/confirmation_dialog.dart';
 import 'package:simple_notes/presentation/widgets/sort_button.dart';
 import 'package:simple_notes/presentation/widgets/shared/app_drawer.dart';
 
@@ -61,7 +62,11 @@ class SkeletonState extends State<Skeleton> {
           IconButton(
             icon: const Icon(Icons.delete),
             onPressed: () {
-              multiselectProvider.deleteSelectedNotes(noteProvider, recicleBinProvider);
+              ConfirmationDialog(context: context, message: 'Do you want to delete all the selected notes?').showConfirmationDialog(context).then((confirmation) {
+                if(confirmation == true){
+                  multiselectProvider.deleteSelectedNotes(noteProvider, recicleBinProvider);
+                }
+              });
             },
           ),
